AvailabilityChecker: Remove hard requirement on tobii_research
Now tobii_research [1][2] module does not need to be installed
to be able to use mouse emulation.

[1]: https://developer.tobiipro.com/python.html
[2]: https://pypi.org/project/tobii-research/

This change removes 'import tobii_research as tr' from
checkPythonEnvironment() method, and adjusts checkEyeTracker() and
getEyeTrackerName() to work correctly even if tobii_research is not
installed.

Not tested yet.

The problem with requiring tobii_research is that it requires specific
Python version to be able to be installed: 3.8 or 3.10 [3][4][5]

[3]: https://www.tobii.com/products/software/applications-and-developer-kits/tobii-pro-sdk
[4]: https://developer.tobiipro.com/tobiiprosdk/platform-and-language.html
[5]: https://codegrits.github.io/CodeGRITS/usage-guide/#python-environment